tested adding user function

+++ JsonStorage.py
@@ -20,19 +20,40 @@ class JsonStorage(Storage, ABC):
         if not os.path.exists(self.base_folder):
             os.makedirs(self.base_folder)
 
-            # Create a JSON file for usernames inside the test folder
-            self.users_file = os.path.join(self.base_folder, 'users.json')
+        # Create a JSON file for usernames inside the test folder
+        self.users_file = os.path.join(self.base_folder, "users.json")
+        if not os.path.exists(self.users_file):
+            with open(self.users_file, 'w') as f:
+                json.dump({"users": []}, f)
 
+
+    # TODO: test if the user is invalid
+    # todo: test that the user folder exists
+    # if it is invalid, raise an error
     def get_user_folder(self, user_id: str):
         user_folder = os.path.join(self.base_folder, user_id)
         if not os.path.exists(user_folder):
-            os.makedirs(user_folder)
+            raise FileNotFoundError(f"User '{user_id}' does not exist")
         return user_folder
 
+    # test that the user folder is created
+    # test the folder wased created before
+    def create_user_folder(self, user_id: str):
+        user_folder = os.path.join(self.base_folder, user_id)
+        if os.path.exists(user_folder):
+            raise FileExistsError(f"User '{user_id}' already exists")
+        os.makedirs(user_folder)
+        return user_folder
+
+    # test that he user is not already in the storage
+    # test that the user is added to the user file
     def add_user_to_storage(self, user_id: str):
-        user_folder = self.get_user_folder(user_id)  # Ensure the user's folder is created
+        if os.path.exists(os.path.join(self.base_folder, user_id)):
+            raise FileExistsError(f"User '{user_id}' already exists")
+        user_folder = self.create_user_folder(user_id)  # Ensure the user's folder is created
 
         # Create a JSON file for the names of collections
+        # I dont think this is needed as we can get the names from the files in the folder
         collections_file = os.path.join(user_folder, 'collections.json')
         with open(collections_file, 'w') as f:
             json.dump([], f)  # Initialize with an empty list
@@ -47,12 +68,29 @@ class JsonStorage(Storage, ABC):
         with open(read_file, 'w') as f:
             json.dump({"name": "read", "books": []}, f)  # Initialize with the correct format
 
+        # Add the user to the users file
+        with open(self.users_file, 'r') as f:
+            users_data = json.load(f)
+        if user_id not in users_data["users"]:
+            users_data["users"].append(user_id)
+            with open(self.users_file, 'w') as f:
+                json.dump(users_data, f, indent=4)
+
+
+
+
+
+
+    # TODO: needs to be tested
+    # TODO: test that the names can be used to get collections from files
+    # TODO: if invalid user, raise an error
     def get_list_of_collection_names(self, user_id: str) -> list:
         user_folder = self.get_user_folder(user_id)
         collection_files = [f for f in os.listdir(user_folder) if f.endswith('.json')]
         collection_names = [os.path.splitext(f)[0] for f in collection_files]
         return collection_names
 
+    # TODO: test that the user is valid
     def remove_user_from_storage(self, user_id: str):
         user_folder = self.get_user_folder(user_id)
 
@@ -62,6 +100,8 @@ class JsonStorage(Storage, ABC):
                 os.remove(file_path)
             os.rmdir(user_folder)
 
+    # TODO: test that the book gets added to collection and books, even if no collection is specified
+    # TODO: test that the book is not added to the collection if it is already there
     def add_book_to_storage(self, book: Book, user_id: str,
                             collection_name="books"):  # May be redundant as all books are in collections
         user_folder = self.get_user_folder(user_id)
@@ -79,10 +119,9 @@ class JsonStorage(Storage, ABC):
         except json.JSONDecodeError as e:
             raise ValueError(f"Error decoding JSON for collection '{collection_name}': {e}")
 
-
         # Ensure collection_data is a list of dictionaries
         if not isinstance(collection_data["books"], list):
-             raise ValueError("Collection data should be a list, the type is: " + str(type(collection_data["books"])))
+            raise ValueError("Collection data should be a list, the type is: " + str(type(collection_data["books"])))
 
         # Check if the book is already in the collection
         if book.to_dict() not in collection_data["books"]:
@@ -90,7 +129,9 @@ class JsonStorage(Storage, ABC):
             with open(collection_path, 'w') as f:
                 json.dump(collection_data, f, indent=4)
 
-
+    # TODO: test the book is removed from the collections and the books
+    # TODO: test error is raised if the book is not in the collection
+    # TODO: test that error is raised is collection path is invalid
     def remove_book_from_storage(self, book: Book, user_id: str, collection_name='books',
                                  remove_from_all_collections=False):  # Will be useful
         user_folder = self.get_user_folder(user_id)
@@ -139,6 +180,9 @@ class JsonStorage(Storage, ABC):
         except json.JSONDecodeError as e:
             raise ValueError(f"Error decoding JSON for collection '{collection_name}': {e}")
 
+    # TODO: test that the collection is added to the collections folder
+    # TODO: test that error is raised if the collection already exists
+    # TODO: test collection can be converted back to a collection object
     def add_collection_to_storage(self, collection: BookCollection, user_id: str):
         user_folder = self.get_user_folder(user_id)
         collection_path = os.path.join(user_folder, f"{collection.name}.json")
@@ -155,6 +199,8 @@ class JsonStorage(Storage, ABC):
         except IOError as e:
             raise IOError(f"Error saving collection '{collection.name}' to storage: {e}")
 
+    # TODO: test that the collection is removed from the user folder
+    # TODO: test that error is raised if the collection does not exist
     def remove_collection_from_storage(self, collection_name: str, user_id: str):
         user_folder = self.get_user_folder(user_id)
         collection_file = f"{collection_name}.json"
@@ -163,6 +209,10 @@ class JsonStorage(Storage, ABC):
         if os.path.exists(collection_path):
             os.remove(collection_path)
 
+    # TODO: test that the collection is loaded correctly into a collection object
+    # TODO: test that error is raised if the collection does not exist
+    # TODO: test that error is raised if the JSON data is invalid
+    # TODO: test that the books are the same
     def load_collection_from_storage(self, user_id: str, collection_name: str) -> BookCollection:
         user_folder = self.get_user_folder(user_id)
         collection_path = os.path.join(user_folder, f"{collection_name}.json")
